+++ src/python/grpcio_channelz/MANIFEST.in
@@ -1,3 +1,4 @@
 include grpc_version.py
 recursive-include grpc_channelz *.py
 global-exclude *.pyc
+include LICENSE
