+++ src/python/grpcio_channelz/channelz_commands.py
@@ -21,10 +21,12 @@ import setuptools
 ROOT_DIR = os.path.abspath(os.path.dirname(os.path.abspath(__file__)))
 CHANNELZ_PROTO = os.path.join(ROOT_DIR,
                               '../../proto/grpc/channelz/channelz.proto')
+LICENSE = os.path.join(ROOT_DIR, '../../../LICENSE')
 
 
-class CopyProtoModules(setuptools.Command):
-    """Command to copy proto modules from grpc/src/proto."""
+class Preprocess(setuptools.Command):
+    """Command to copy proto modules from grpc/src/proto and LICENSE from
+    the root directory"""
 
     description = ''
     user_options = []
@@ -40,6 +42,8 @@ class CopyProtoModules(setuptools.Command):
             shutil.copyfile(CHANNELZ_PROTO,
                             os.path.join(ROOT_DIR,
                                          'grpc_channelz/v1/channelz.proto'))
+        if os.path.isfile(LICENSE):
+            shutil.copyfile(LICENSE, os.path.join(ROOT_DIR, 'LICENSE'))
 
 
 class BuildPackageProtos(setuptools.Command):
