+++ src/python/grpcio_health_checking/MANIFEST.in
@@ -1,3 +1,4 @@
 include grpc_version.py
 recursive-include grpc_health *.py
 global-exclude *.pyc
+include LICENSE
