+++ src/python/grpcio_reflection/MANIFEST.in
@@ -1,3 +1,4 @@
 include grpc_version.py
 recursive-include grpc_reflection *.py
 global-exclude *.pyc
+include LICENSE
