+++ src/python/grpcio_testing/MANIFEST.in
@@ -1,3 +1,4 @@
 include grpc_version.py
 recursive-include grpc_testing *.py
 global-exclude *.pyc
+include LICENSE
